# Conflicts:
#	src/main/java/Controllers/CategoryController.java
#	src/main/java/Controllers/CustomerController.java
#	src/main/java/Controllers/ManagerController.java
#	src/main/java/Controllers/NewManagerController.java
#	src/main/java/Controllers/ProductController.java
#	src/main/java/Models/Category.java
#	src/main/java/Models/ConfirmationStatus.java
#	src/main/java/Models/Customer.java
#	src/main/java/Models/Field.java
#	src/main/java/Models/Item.java
#	src/main/java/Models/Log.java
#	src/main/java/Models/Pendable.java
#	src/main/java/Models/Sale.java
#	src/main/java/View/EntryMenu.java
#	src/main/java/View/ManageAllProductsMenu.java
#	src/main/java/View/ManageCategoriesMenu.java
#	src/main/java/View/ManageRequestsMenu.java
#	src/main/java/View/ManageUsersMenu.java
#	src/main/java/View/ManagerMenu.java
#	src/main/java/View/Menu.java
#	src/main/java/View/ViewDiscountCodesMenu.java

+++ src/main/java/Models/Score.java
@@ -2,9 +2,9 @@ package Models;
 
 public class Score {
     private Customer customer;
-    private double score;
+    private int score;
 
-    public Score(Customer customer, double score) {
+    public Score(Customer customer, int score) {
         this.customer = customer;
         this.score = score;
     }
@@ -13,7 +13,7 @@ public class Score {
         return customer;
     }
 
-    public double getScore() {
+    public int getScore() {
         return score;
     }
